CASMPET-7553: If helm status is 'uninstalled' return success in undeploy function. CASMPET-7552: In upgrade_control_plane.sh, use manifests in CSM_ARTI_DIR/build/manifests directory that have gone through manifestgen.

+++ upgrade/scripts/k8s/upgrade_control_plane.sh
@@ -165,8 +165,13 @@ function deploy() {
 # Use this if a chart has been removed from a manifest and needs
 # to be removed from the system as part of an upgrade.
 function undeploy() {
-  # If the chart is missing (rc==1) just return success.
+  # Now that we're using --keep-history, helm status will return with a STATUS
+  # of "uninstalled" if the chart has already been uninstalled.
+  # If the chart is missing (rc==1) or if uninstalled, return success.
   helm status "$@" || return 0
+  if [ "$(helm status "$@" | grep STATUS | awk '{print $2}')" = "uninstalled" ]; then
+    return 0
+  fi
   # Remove the chart.
   helm uninstall "$@" --keep-history
 }
@@ -178,7 +183,7 @@ if [ ${k8s_minor_version} -gt 24 ]; then
 fi
 
 # If there are post-upgrade-*-<k8s_version>.yaml files, deploy charts in those files.
-manifests_dir="${CSM_ARTI_DIR}/manifests"
+manifests_dir="${CSM_ARTI_DIR}/build/manifests"
 find "${manifests_dir}/" -name "post-upgrade-*-${k8s_version}.yaml" | sort | while read -r manifest; do
   echo "INFO Deploying ${manifest} ..."
   deploy "${manifest}"
